requirements.txt
flask==2.3.3
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.9.0.80
numpy==1.26.4
sqlite3
pillow==10.3.0
gunicorn==22.0.0